requirements.txt
Flask==2.0.3
Flask-MySQLdb==1.0.0  # For MySQL database connection (if using MySQL)
Flask-SQLAlchemy==2.5.1  # Optional: for ORM and easier database interaction
Flask-WTF==1.0.0  # Optional: for form handling and validation
